Centralize ArrayNode logic in one file
