+++ packages/dds/tree/src/simple-tree/types.ts
@@ -3,10 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { FlexTreeNodeSchema } from "../feature-libraries/index.js";
-
 import { WithType, type } from "./schemaTypes.js";
-import { IterableTreeArrayContent } from "./treeArrayNode.js";
 
 /**
  * Type alias to document which values are un-hydrated.
@@ -71,167 +68,3 @@ export abstract class TreeNode implements WithType {
 	 */
 	public abstract get [type](): string;
 }
-
-/**
- * A generic array type, used to defined types like {@link (TreeArrayNode:interface)}.
- *
- * @privateRemarks
- * Inlining this into TreeArrayNode causes recursive array use to stop compiling.
- *
- * @public
- */
-export interface TreeArrayNodeBase<out T, in TNew, in TMoveFrom>
-	extends ReadonlyArray<T>,
-		TreeNode {
-	/**
-	 * Inserts new item(s) at a specified location.
-	 * @param index - The index at which to insert `value`.
-	 * @param value - The content to insert.
-	 * @throws Throws if `index` is not in the range [0, `array.length`).
-	 */
-	insertAt(index: number, ...value: (TNew | IterableTreeArrayContent<TNew>)[]): void;
-
-	/**
-	 * Inserts new item(s) at the start of the array.
-	 * @param value - The content to insert.
-	 */
-	insertAtStart(...value: (TNew | IterableTreeArrayContent<TNew>)[]): void;
-
-	/**
-	 * Inserts new item(s) at the end of the array.
-	 * @param value - The content to insert.
-	 */
-	insertAtEnd(...value: (TNew | IterableTreeArrayContent<TNew>)[]): void;
-
-	/**
-	 * Removes the item at the specified location.
-	 * @param index - The index at which to remove the item.
-	 * @throws Throws if `index` is not in the range [0, `array.length`).
-	 */
-	removeAt(index: number): void;
-
-	/**
-	 * Removes all items between the specified indices.
-	 * @param start - The starting index of the range to remove (inclusive). Defaults to the start of the array.
-	 * @param end - The ending index of the range to remove (exclusive).
-	 * @throws Throws if `start` is not in the range [0, `array.length`).
-	 * @throws Throws if `end` is less than `start`.
-	 * If `end` is not supplied or is greater than the length of the array, all items after `start` are removed.
-	 */
-	removeRange(start?: number, end?: number): void;
-
-	/**
-	 * Moves the specified item to the start of the array.
-	 * @param sourceIndex - The index of the item to move.
-	 * @throws Throws if `sourceIndex` is not in the range [0, `array.length`).
-	 */
-	moveToStart(sourceIndex: number): void;
-
-	/**
-	 * Moves the specified item to the start of the array.
-	 * @param sourceIndex - The index of the item to move.
-	 * @param source - The source array to move the item out of.
-	 * @throws Throws if `sourceIndex` is not in the range [0, `array.length`).
-	 */
-	moveToStart(sourceIndex: number, source: TMoveFrom): void;
-
-	/**
-	 * Moves the specified item to the end of the array.
-	 * @param sourceIndex - The index of the item to move.
-	 * @throws Throws if `sourceIndex` is not in the range [0, `array.length`).
-	 */
-	moveToEnd(sourceIndex: number): void;
-
-	/**
-	 * Moves the specified item to the end of the array.
-	 * @param sourceIndex - The index of the item to move.
-	 * @param source - The source array to move the item out of.
-	 * @throws Throws if `sourceIndex` is not in the range [0, `array.length`).
-	 */
-	moveToEnd(sourceIndex: number, source: TMoveFrom): void;
-
-	/**
-	 * Moves the specified item to the desired location in the array.
-	 * @param index - The index to move the item to.
-	 * This is based on the state of the array before moving the source item.
-	 * @param sourceIndex - The index of the item to move.
-	 * @throws Throws if any of the input indices are not in the range [0, `array.length`).
-	 */
-	moveToIndex(index: number, sourceIndex: number): void;
-
-	/**
-	 * Moves the specified item to the desired location in the array.
-	 * @param index - The index to move the item to.
-	 * @param sourceIndex - The index of the item to move.
-	 * @param source - The source array to move the item out of.
-	 * @throws Throws if any of the input indices are not in the range [0, `array.length`).
-	 */
-	moveToIndex(index: number, sourceIndex: number, source: TMoveFrom): void;
-
-	/**
-	 * Moves the specified items to the start of the array.
-	 * @param sourceStart - The starting index of the range to move (inclusive).
-	 * @param sourceEnd - The ending index of the range to move (exclusive)
-	 * @throws Throws if either of the input indices are not in the range [0, `array.length`) or if `sourceStart` is greater than `sourceEnd`.
-	 */
-	moveRangeToStart(sourceStart: number, sourceEnd: number): void;
-
-	/**
-	 * Moves the specified items to the start of the array.
-	 * @param sourceStart - The starting index of the range to move (inclusive).
-	 * @param sourceEnd - The ending index of the range to move (exclusive)
-	 * @param source - The source array to move items out of.
-	 * @throws Throws if the types of any of the items being moved are not allowed in the destination array,
-	 * if either of the input indices are not in the range [0, `array.length`) or if `sourceStart` is greater than `sourceEnd`.
-	 */
-	moveRangeToStart(sourceStart: number, sourceEnd: number, source: TMoveFrom): void;
-
-	/**
-	 * Moves the specified items to the end of the array.
-	 * @param sourceStart - The starting index of the range to move (inclusive).
-	 * @param sourceEnd - The ending index of the range to move (exclusive)
-	 * @throws Throws if either of the input indices are not in the range [0, `array.length`) or if `sourceStart` is greater than `sourceEnd`.
-	 */
-	moveRangeToEnd(sourceStart: number, sourceEnd: number): void;
-
-	/**
-	 * Moves the specified items to the end of the array.
-	 * @param sourceStart - The starting index of the range to move (inclusive).
-	 * @param sourceEnd - The ending index of the range to move (exclusive)
-	 * @param source - The source array to move items out of.
-	 * @throws Throws if the types of any of the items being moved are not allowed in the destination array,
-	 * if either of the input indices are not in the range [0, `array.length`) or if `sourceStart` is greater than `sourceEnd`.
-	 */
-	moveRangeToEnd(sourceStart: number, sourceEnd: number, source: TMoveFrom): void;
-
-	/**
-	 * Moves the specified items to the desired location within the array.
-	 * @param index - The index to move the items to.
-	 * This is based on the state of the array before moving the source items.
-	 * @param sourceStart - The starting index of the range to move (inclusive).
-	 * @param sourceEnd - The ending index of the range to move (exclusive)
-	 * @throws Throws if any of the input indices are not in the range [0, `array.length`) or if `sourceStart` is greater than `sourceEnd`.
-	 */
-	moveRangeToIndex(index: number, sourceStart: number, sourceEnd: number): void;
-
-	/**
-	 * Moves the specified items to the desired location within the array.
-	 * @param index - The index to move the items to.
-	 * @param sourceStart - The starting index of the range to move (inclusive).
-	 * @param sourceEnd - The ending index of the range to move (exclusive)
-	 * @param source - The source array to move items out of.
-	 * @throws Throws if the types of any of the items being moved are not allowed in the destination array,
-	 * if any of the input indices are not in the range [0, `array.length`) or if `sourceStart` is greater than `sourceEnd`.
-	 */
-	moveRangeToIndex(
-		index: number,
-		sourceStart: number,
-		sourceEnd: number,
-		source: TMoveFrom,
-	): void;
-}
-
-/**
- * Given a node's schema, return the corresponding object in the proxy-based API.
- */
-export type TypedNode<TSchema extends FlexTreeNodeSchema> = TreeNode & WithType<TSchema["name"]>;
